backend/requirements.txt
Django==5.2.4
django-ninja==1.1.0
django-cors-headers==4.3.1
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.1
pyjwt==2.10.1
faker==37.5.3